Fix build errors on NixOS

+++ boot/efi/kernel_impl.c
@@ -269,7 +269,7 @@ int netdrv_send(NetBoot_Device device, void* buffer, int size) {
 }
 
 
-void memcpy(void* dst, const void* src, int size) {
+void memcpy(void* dst, const void* src, size_t size) {
     if (dst == src)
         return;
     for (int i=0;i<size;i++) {
